push of latest code tested on track, in cup mode

+++ sensorcheck - Copy/Rover.h
@@ -42,7 +42,7 @@ class Rover
         );
 
         void begin();
-        void calibrate(unsigned long int min[3], unsigned long int max[3]);
+        void calibrate(int min[3], int max[3]);
 
         //steering controls
         void steerLeft(int degrees);
@@ -71,8 +71,8 @@ class Rover
         Pixy2 pixy;
     
     private:
-        unsigned long int _min[3];
-        unsigned long int _max[3];
+        float _min[3];
+        float _max[3];
         int _STEERING_SERVO;
         int _CLAW_SERVO;
         int _MOTOR;
